refactor: change zone ID from u64/uint64 (8 bytes) to u32/uint32 (4 bytes)
Narrows zone ID to 4 bytes so it packs more easily with other fields.
Applies across Solidity contracts, Rust types, RPC auth token binary
format, xtask tooling, docs, and the Justfile.

Auth token fields shrink from 53 to 49 bytes (1+4+8+20+8+8).

Amp-Thread-ID: https://ampcode.com/threads/T-019cf563-8a24-714b-9419-dd231ff176ef

+++ crates/primitives/src/abi.rs
@@ -170,7 +170,7 @@ macro_rules! define_abi {
 
                 // -- View functions --
 
-                function zoneId() external view returns (uint64);
+                function zoneId() external view returns (uint32);
                 function sequencer() external view returns (address);
                 function verifier() external view returns (address);
                 function sequencerPubkey() external view returns (bytes32);
@@ -369,7 +369,7 @@ macro_rules! define_abi {
 
             #[derive(Debug)]
             struct ZoneInfo {
-                uint64 zoneId;
+                uint32 zoneId;
                 address portal;
                 address messenger;
                 address initialToken;
@@ -395,7 +395,7 @@ macro_rules! define_abi {
                 }
                 #[derive(Debug)]
                 event ZoneCreated(
-                    uint64 indexed zoneId,
+                    uint32 indexed zoneId,
                     address indexed portal,
                     address indexed messenger,
                     address token,
@@ -405,10 +405,10 @@ macro_rules! define_abi {
                     bytes32 genesisTempoBlockHash,
                     uint64 genesisTempoBlockNumber
                 );
-                function createZone(CreateZoneParams calldata params) external returns (uint64 zoneId, address portal);
+                function createZone(CreateZoneParams calldata params) external returns (uint32 zoneId, address portal);
                 function verifier() external view returns (address);
-                function zones(uint64 zoneId) external view returns (ZoneInfo memory);
-                function zoneCount() external view returns (uint64);
+                function zones(uint32 zoneId) external view returns (ZoneInfo memory);
+                function zoneCount() external view returns (uint32);
                 function isZonePortal(address portal) external view returns (bool);
                 function isZoneMessenger(address messenger) external view returns (bool);
             }
